10/20

+++ Unit 2 Project/Models/Employee.cs
@@ -13,6 +13,6 @@
         public string? Position { get; set; }
 
         // Employee Salary
-        public double? Salary { get; set; }
+        public decimal? Salary { get; set; }
     }
 }
